Refactored retirement trends class names
