--- carbonmark-data/components/cards/retirementTrends/KlimaDAORetirementsByChainOverviewCard/index.tsx
@@ -1,22 +1,0 @@
-import { t } from "@lingui/macro";
-import ChartCard, { CardProps } from "components/cards/ChartCard";
-import KlimaDAORetirementsByChainOverviewChart from "components/charts/KlimaDAORetirementsByChainOverviewChart";
-
-/** Klima DAO Retirements by token Card */
-export default function KlimaDAORetirementsByChainOverviewCard(
-  props: CardProps
-) {
-  const chart = (
-    /* @ts-expect-error async Server component */
-    <KlimaDAORetirementsByChainOverviewChart />
-  );
-
-  return (
-    <ChartCard
-      {...props}
-      isColumnCard={true}
-      title={t`Carbon retirements`}
-      chart={chart}
-    />
-  );
-}
